refactor(map): make SystemClass the single source of the wormhole class list
The map server's connection scoping and the new kill notifications each carried
their own copy of the wormhole class ids. Two literal lists of the same thirteen
numbers is exactly what review flags: they can drift, and the drift is silent -
a class added to one and not the other changes which connections get scoped
without changing which kills get posted.

@wh_space in ConnectionsImpl is now derived from
WandererApp.SystemClass.wormhole_classes/0 rather than restated. The six class
attributes that became unused are commented out rather than deleted, matching
the file's own convention for the ids it keeps as a reference table (@ccp2,
@zarzakh). @c13 stays live because it is still referenced elsewhere.

The module also moves out of the Kills namespace to WandererApp.SystemClass. It
was introduced in this branch, so the rename costs nothing, and the map server
depending on a module under Kills would have been the wrong direction.
